Added if statement to the code generator.
If statement code generation essentially works like this:

if cond == true, go to then_label, else go to else_label

then_label:
  ...
  ...
  body of if statement
  ...
  ...

else_label:
  ...
  ...
  body of the next if statement
  ...
  ...

where the else_label also contains a comparison instruction, in order
to move to the next (if any) if statement, assuming the condition
is false for else_label.

Eventually, the results of running any of the if statement bodies
will be in the returned id of the function gen(n: Cond, ...)
when called.